Legg utledGrunnlagTidslinjePerPerson og underfunskjoner inn i GrunnlagForVedtaksperioder
